Diagrams.Color.XKCD

Description

The 949 most common RGB monitor colors, as determined by the xkcd color name survey. See http://xkcd.com/color/rgb/ . Data taken from http://xkcd.com/color/rgb.txt (License: http://creativecommons.org/publicdomain/zero/1.0/ .)

This module defines many common color names that no standards committee would ever countenance, including crude ones like poopGreen and puke, as well as more acceptable but still non-standards-worthy names like windowsBlue and toxicGreen. It also contains several misspellings such as liliac and manilla, though by definition this list contains only very common misspellings.

Names in the original list that use a forward slash (e.g. "yellow/green") are represented here with an underscore, e.g. yellow_green (note that yellow_green /= yellowGreen!).

Note that the color name tan clashes with the Haskell Prelude. Many other color names here clash with names from Data.Colour.Names (taken from the X11 standard), which is automatically re-exported from Diagrams.Prelude. It is recommended to import this module qualified.

Documentation

cloudyBlue :: AlphaColour Double

darkPastelGreen :: AlphaColour Double

dust :: AlphaColour Double

electricLime :: AlphaColour Double

freshGreen :: AlphaColour Double

lightEggplant :: AlphaColour Double

nastyGreen :: AlphaColour Double

reallyLightBlue :: AlphaColour Double

tea :: AlphaColour Double

warmPurple :: AlphaColour Double

yellowishTan :: AlphaColour Double

cement :: AlphaColour Double

darkGrassGreen :: AlphaColour Double

dustyTeal :: AlphaColour Double

greyTeal :: AlphaColour Double

macaroniAndCheese :: AlphaColour Double

pinkishTan :: AlphaColour Double

spruce :: AlphaColour Double

strongBlue :: AlphaColour Double

toxicGreen :: AlphaColour Double

windowsBlue :: AlphaColour Double

blueBlue :: AlphaColour Double

blueWithAHintOfPurple :: AlphaColour Double

booger :: AlphaColour Double

brightSeaGreen :: AlphaColour Double

darkGreenBlue :: AlphaColour Double

deepTurquoise :: AlphaColour Double

greenTeal :: AlphaColour Double

strongPink :: AlphaColour Double

bland :: AlphaColour Double

deepAqua :: AlphaColour Double

lavenderPink :: AlphaColour Double

lightMossGreen :: AlphaColour Double

lightSeafoamGreen :: AlphaColour Double

oliveYellow :: AlphaColour Double

pigPink :: AlphaColour Double

deepLilac :: AlphaColour Double

desert :: AlphaColour Double

dustyLavender :: AlphaColour Double

purpleyGrey :: AlphaColour Double

purply :: AlphaColour Double

candyPink :: AlphaColour Double

lightPastelGreen :: AlphaColour Double

boringGreen :: AlphaColour Double

kiwiGreen :: AlphaColour Double

lightGreyGreen :: AlphaColour Double

orangePink :: AlphaColour Double

teaGreen :: AlphaColour Double

veryLightBrown :: AlphaColour Double

eggShell :: AlphaColour Double

eggplantPurple :: AlphaColour Double

powderPink :: AlphaColour Double

reddishGrey :: AlphaColour Double

babyShitBrown :: AlphaColour Double

liliac :: AlphaColour Double

stormyBlue :: AlphaColour Double

uglyBrown :: AlphaColour Double

custard :: AlphaColour Double

darkishPink :: AlphaColour Double

deepBrown :: AlphaColour Double

greenishBeige :: AlphaColour Double

manilla :: AlphaColour Double

offBlue :: AlphaColour Double

battleshipGrey :: AlphaColour Double

brownyGreen :: AlphaColour Double

bruise :: AlphaColour Double

kelleyGreen :: AlphaColour Double

sicklyYellow :: AlphaColour Double

sunnyYellow :: AlphaColour Double

azul :: AlphaColour Double

darkgreen :: AlphaColour Double

green_yellow :: AlphaColour Double

lichen :: AlphaColour Double

lightLightGreen :: AlphaColour Double

paleGold :: AlphaColour Double

sunYellow :: AlphaColour Double

tanGreen :: AlphaColour Double

burple :: AlphaColour Double

butterscotch :: AlphaColour Double

toupe :: AlphaColour Double

darkCream :: AlphaColour Double

indianRed :: AlphaColour Double

lightLavendar :: AlphaColour Double

poisonGreen :: AlphaColour Double

babyPukeGreen :: AlphaColour Double

brightYellowGreen :: AlphaColour Double

charcoalGrey :: AlphaColour Double

squash :: AlphaColour Double

cinnamon :: AlphaColour Double

lightPeaGreen :: AlphaColour Double

radioactiveGreen :: AlphaColour Double

rawSienna :: AlphaColour Double

babyPurple :: AlphaColour Double

cocoa :: AlphaColour Double

lightRoyalBlue :: AlphaColour Double

orangeish :: AlphaColour Double

rustBrown :: AlphaColour Double

sandBrown :: AlphaColour Double

swamp :: AlphaColour Double

tealishGreen :: AlphaColour Double

burntSiena :: AlphaColour Double

camo :: AlphaColour Double

duskBlue :: AlphaColour Double

fern :: AlphaColour Double

oldRose :: AlphaColour Double

paleLightGreen :: AlphaColour Double

peachyPink :: AlphaColour Double

rosyPink :: AlphaColour Double

lightBluishGreen :: AlphaColour Double

lightBrightGreen :: AlphaColour Double

lightNeonGreen :: AlphaColour Double

lightSeafoam :: AlphaColour Double

tiffanyBlue :: AlphaColour Double

washedOutGreen :: AlphaColour Double

brownyOrange :: AlphaColour Double

niceBlue :: AlphaColour Double

sapphire :: AlphaColour Double

greyishTeal :: AlphaColour Double

orangeyYellow :: AlphaColour Double

parchment :: AlphaColour Double

straw :: AlphaColour Double

veryDarkBrown :: AlphaColour Double

terracota :: AlphaColour Double

uglyBlue :: AlphaColour Double

clearBlue :: AlphaColour Double

creme :: AlphaColour Double

foamGreen :: AlphaColour Double

grey_green :: AlphaColour Double

lightGold :: AlphaColour Double

seafoamBlue :: AlphaColour Double

topaz :: AlphaColour Double

violetPink :: AlphaColour Double

wintergreen :: AlphaColour Double

yellowTan :: AlphaColour Double

darkFuchsia :: AlphaColour Double

indigoBlue :: AlphaColour Double

lightYellowishGreen :: AlphaColour Double

paleMagenta :: AlphaColour Double

richPurple :: AlphaColour Double

sunflowerYellow :: AlphaColour Double

green_blue :: AlphaColour Double

leather :: AlphaColour Double

racingGreen :: AlphaColour Double

vividPurple :: AlphaColour Double

darkRoyalBlue :: AlphaColour Double

hazel :: AlphaColour Double

mutedPink :: AlphaColour Double

boogerGreen :: AlphaColour Double

canary :: AlphaColour Double

coolGrey :: AlphaColour Double

darkTaupe :: AlphaColour Double

darkishPurple :: AlphaColour Double

trueGreen :: AlphaColour Double

coralPink :: AlphaColour Double

darkSage :: AlphaColour Double

darkSlateBlue :: AlphaColour Double

flatBlue :: AlphaColour Double

mushroom :: AlphaColour Double

richBlue :: AlphaColour Double

dirtyPurple :: AlphaColour Double

greenblue :: AlphaColour Double

ickyGreen :: AlphaColour Double

lightKhaki :: AlphaColour Double

warmBlue :: AlphaColour Double

darkHotPink :: AlphaColour Double

deepSeaBlue :: AlphaColour Double

carmine :: AlphaColour Double

darkYellowGreen :: AlphaColour Double

palePeach :: AlphaColour Double

plumPurple :: AlphaColour Double

goldenRod :: AlphaColour Double

neonRed :: AlphaColour Double

oldPink :: AlphaColour Double

veryPaleBlue :: AlphaColour Double

bloodOrange :: AlphaColour Double

grapefruit :: AlphaColour Double

sandYellow :: AlphaColour Double

clayBrown :: AlphaColour Double

darkBlueGrey :: AlphaColour Double

flatGreen :: AlphaColour Double

lightGreenBlue :: AlphaColour Double

warmPink :: AlphaColour Double

dodgerBlue :: AlphaColour Double

grossGreen :: AlphaColour Double

ice :: AlphaColour Double

metallicBlue :: AlphaColour Double

paleSalmon :: AlphaColour Double

sapGreen :: AlphaColour Double

algae :: AlphaColour Double

blueyGrey :: AlphaColour Double

greenyGrey :: AlphaColour Double

highlighterGreen :: AlphaColour Double

lightLightBlue :: AlphaColour Double

lightMint :: AlphaColour Double

rawUmber :: AlphaColour Double

vividBlue :: AlphaColour Double

deepLavender :: AlphaColour Double

dullTeal :: AlphaColour Double

lightGreenishBlue :: AlphaColour Double

mudGreen :: AlphaColour Double

pinky :: AlphaColour Double

redWine :: AlphaColour Double

shitGreen :: AlphaColour Double

tanBrown :: AlphaColour Double

darkblue :: AlphaColour Double

rosa :: AlphaColour Double

lipstick :: AlphaColour Double

paleMauve :: AlphaColour Double

claret :: AlphaColour Double

dandelion :: AlphaColour Double

orangered :: AlphaColour Double

poopGreen :: AlphaColour Double

ruby :: AlphaColour Double

dark :: AlphaColour Double

greenishTurquoise :: AlphaColour Double

pastelRed :: AlphaColour Double

pissYellow :: AlphaColour Double

brightCyan :: AlphaColour Double

darkCoral :: AlphaColour Double

algaeGreen :: AlphaColour Double

darkishRed :: AlphaColour Double

reddyBrown :: AlphaColour Double

blushPink :: AlphaColour Double

camouflageGreen :: AlphaColour Double

lawnGreen :: AlphaColour Double

putty :: AlphaColour Double

vibrantBlue :: AlphaColour Double

darkSand :: AlphaColour Double

purple_blue :: AlphaColour Double

saffron :: AlphaColour Double

twilight :: AlphaColour Double

warmBrown :: AlphaColour Double

bluegrey :: AlphaColour Double

bubbleGumPink :: AlphaColour Double

duckEggBlue :: AlphaColour Double

greenishCyan :: AlphaColour Double

petrol :: AlphaColour Double

royal :: AlphaColour Double

butter :: AlphaColour Double

dustyOrange :: AlphaColour Double

offYellow :: AlphaColour Double

paleOliveGreen :: AlphaColour Double

orangish :: AlphaColour Double

leaf :: AlphaColour Double

lightBlueGrey :: AlphaColour Double

driedBlood :: AlphaColour Double

lightishPurple :: AlphaColour Double

rustyRed :: AlphaColour Double

lavenderBlue :: AlphaColour Double

lightGrassGreen :: AlphaColour Double

lightMintGreen :: AlphaColour Double

sunflower :: AlphaColour Double

velvet :: AlphaColour Double

brickOrange :: AlphaColour Double

lightishRed :: AlphaColour Double

pureBlue :: AlphaColour Double

twilightBlue :: AlphaColour Double

violetRed :: AlphaColour Double

yellowyBrown :: AlphaColour Double

carnation :: AlphaColour Double

muddyYellow :: AlphaColour Double

darkSeafoamGreen :: AlphaColour Double

deepRose :: AlphaColour Double

dustyRed :: AlphaColour Double

grey_blue :: AlphaColour Double

lemonLime :: AlphaColour Double

purple_pink :: AlphaColour Double

brownYellow :: AlphaColour Double

purpleBrown :: AlphaColour Double

wisteria :: AlphaColour Double

bananaYellow :: AlphaColour Double

lipstickRed :: AlphaColour Double

waterBlue :: AlphaColour Double

brownGrey :: AlphaColour Double

vibrantPurple :: AlphaColour Double

babyGreen :: AlphaColour Double

barfGreen :: AlphaColour Double

eggshellBlue :: AlphaColour Double

sandyYellow :: AlphaColour Double

coolGreen :: AlphaColour Double

pale :: AlphaColour Double

blue_grey :: AlphaColour Double

hotMagenta :: AlphaColour Double

greyblue :: AlphaColour Double

purpley :: AlphaColour Double

babyShitGreen :: AlphaColour Double

brownishPink :: AlphaColour Double

darkAquamarine :: AlphaColour Double

diarrhea :: AlphaColour Double

lightMustard :: AlphaColour Double

paleSkyBlue :: AlphaColour Double

turtleGreen :: AlphaColour Double

brightOlive :: AlphaColour Double

darkGreyBlue :: AlphaColour Double

greenyBrown :: AlphaColour Double

lemonGreen :: AlphaColour Double

lightPeriwinkle :: AlphaColour Double

seaweedGreen :: AlphaColour Double

sunshineYellow :: AlphaColour Double

uglyPurple :: AlphaColour Double

mediumPink :: AlphaColour Double

pukeBrown :: AlphaColour Double

veryLightPink :: AlphaColour Double

viridian :: AlphaColour Double

bile :: AlphaColour Double

fadedYellow :: AlphaColour Double

veryPaleGreen :: AlphaColour Double

vibrantGreen :: AlphaColour Double

brightLime :: AlphaColour Double

spearmint :: AlphaColour Double

lightAquamarine :: AlphaColour Double

lightSage :: AlphaColour Double

yellowgreen :: AlphaColour Double

babyPoo :: AlphaColour Double

darkSeafoam :: AlphaColour Double

deepTeal :: AlphaColour Double

heather :: AlphaColour Double

rustOrange :: AlphaColour Double

dirtyBlue :: AlphaColour Double

fernGreen :: AlphaColour Double

brightLilac :: AlphaColour Double

weirdGreen :: AlphaColour Double

peacockBlue :: AlphaColour Double

avocadoGreen :: AlphaColour Double

fadedOrange :: AlphaColour Double

grapePurple :: AlphaColour Double

hotGreen :: AlphaColour Double

limeYellow :: AlphaColour Double

mango :: AlphaColour Double

shamrock :: AlphaColour Double

bubblegum :: AlphaColour Double

purplishBrown :: AlphaColour Double

vomitYellow :: AlphaColour Double

paleCyan :: AlphaColour Double

keyLime :: AlphaColour Double

tomatoRed :: AlphaColour Double

lightgreen :: AlphaColour Double

merlot :: AlphaColour Double

nightBlue :: AlphaColour Double

purpleishPink :: AlphaColour Double

apple :: AlphaColour Double

babyPoopGreen :: AlphaColour Double

greenApple :: AlphaColour Double

heliotrope :: AlphaColour Double

yellow_green :: AlphaColour Double

almostBlack :: AlphaColour Double

coolBlue :: AlphaColour Double

leafyGreen :: AlphaColour Double

mustardBrown :: AlphaColour Double

dusk :: AlphaColour Double

dullBrown :: AlphaColour Double

frogGreen :: AlphaColour Double

vividGreen :: AlphaColour Double

brightLightGreen :: AlphaColour Double

fluroGreen :: AlphaColour Double

kiwi :: AlphaColour Double

seaweed :: AlphaColour Double

navyGreen :: AlphaColour Double

ultramarineBlue :: AlphaColour Double

iris :: AlphaColour Double

pastelOrange :: AlphaColour Double

yellowishOrange :: AlphaColour Double

perrywinkle :: AlphaColour Double

tealish :: AlphaColour Double

darkPlum :: AlphaColour Double

pear :: AlphaColour Double

pinkishOrange :: AlphaColour Double

midnightPurple :: AlphaColour Double

lightUrple :: AlphaColour Double

darkMint :: AlphaColour Double

greenishTan :: AlphaColour Double

lightBurgundy :: AlphaColour Double

turquoiseBlue :: AlphaColour Double

uglyPink :: AlphaColour Double

sandy :: AlphaColour Double

electricPink :: AlphaColour Double

mutedPurple :: AlphaColour Double

midGreen :: AlphaColour Double

greyish :: AlphaColour Double

neonYellow :: AlphaColour Double

banana :: AlphaColour Double

carnationPink :: AlphaColour Double

tomato :: AlphaColour Double

sea :: AlphaColour Double

muddyBrown :: AlphaColour Double

turquoiseGreen :: AlphaColour Double

buff :: AlphaColour Double

fawn :: AlphaColour Double

mutedBlue :: AlphaColour Double

paleRose :: AlphaColour Double

darkMintGreen :: AlphaColour Double

amethyst :: AlphaColour Double

blue_green :: AlphaColour Double

chestnut :: AlphaColour Double

sickGreen :: AlphaColour Double

pea :: AlphaColour Double

rustyOrange :: AlphaColour Double

stone :: AlphaColour Double

roseRed :: AlphaColour Double

paleAqua :: AlphaColour Double

deepOrange :: AlphaColour Double

earth :: AlphaColour Double

mossyGreen :: AlphaColour Double

grassyGreen :: AlphaColour Double

paleLimeGreen :: AlphaColour Double

lightGreyBlue :: AlphaColour Double

paleGrey :: AlphaColour Double

asparagus :: AlphaColour Double

blueberry :: AlphaColour Double

purpleRed :: AlphaColour Double

paleLime :: AlphaColour Double

greenishTeal :: AlphaColour Double

caramel :: AlphaColour Double

deepMagenta :: AlphaColour Double

lightPeach :: AlphaColour Double

milkChocolate :: AlphaColour Double

ocher :: AlphaColour Double

offGreen :: AlphaColour Double

purplyPink :: AlphaColour Double

lightblue :: AlphaColour Double

duskyBlue :: AlphaColour Double

golden :: AlphaColour Double

lightBeige :: AlphaColour Double

butterYellow :: AlphaColour Double

duskyPurple :: AlphaColour Double

frenchBlue :: AlphaColour Double

uglyYellow :: AlphaColour Double

greenyYellow :: AlphaColour Double

orangishRed :: AlphaColour Double

shamrockGreen :: AlphaColour Double

orangishBrown :: AlphaColour Double

treeGreen :: AlphaColour Double

deepViolet :: AlphaColour Double

gunmetal :: AlphaColour Double

blue_purple :: AlphaColour Double

cherry :: AlphaColour Double

sandyBrown :: AlphaColour Double

warmGrey :: AlphaColour Double

darkIndigo :: AlphaColour Double

midnight :: AlphaColour Double

blueyGreen :: AlphaColour Double

greyPink :: AlphaColour Double

softPurple :: AlphaColour Double

blood :: AlphaColour Double

brownRed :: AlphaColour Double

mediumGrey :: AlphaColour Double

berry :: AlphaColour Double

poo :: AlphaColour Double

purpleyPink :: AlphaColour Double

lightSalmon :: AlphaColour Double

snot :: AlphaColour Double

easterPurple :: AlphaColour Double

lightYellowGreen :: AlphaColour Double

darkNavyBlue :: AlphaColour Double

drab :: AlphaColour Double

lightRose :: AlphaColour Double

rouge :: AlphaColour Double

purplishRed :: AlphaColour Double

slimeGreen :: AlphaColour Double

babyPoop :: AlphaColour Double

irishGreen :: AlphaColour Double

pink_purple :: AlphaColour Double

darkNavy :: AlphaColour Double

greenyBlue :: AlphaColour Double

lightPlum :: AlphaColour Double

pinkishGrey :: AlphaColour Double

dirtyOrange :: AlphaColour Double

rustRed :: AlphaColour Double

paleLilac :: AlphaColour Double

orangeyRed :: AlphaColour Double

primaryBlue :: AlphaColour Double

kermitGreen :: AlphaColour Double

brownishPurple :: AlphaColour Double

murkyGreen :: AlphaColour Double

wheat :: AlphaColour Double

veryDarkPurple :: AlphaColour Double

bottleGreen :: AlphaColour Double

watermelon :: AlphaColour Double

deepSkyBlue :: AlphaColour Double

fireEngineRed :: AlphaColour Double

yellowOchre :: AlphaColour Double

pumpkinOrange :: AlphaColour Double

paleOlive :: AlphaColour Double

lightLilac :: AlphaColour Double

lightishGreen :: AlphaColour Double

carolinaBlue :: AlphaColour Double

mulberry :: AlphaColour Double

shockingPink :: AlphaColour Double

auburn :: AlphaColour Double

brightLimeGreen :: AlphaColour Double

celadon :: AlphaColour Double

pinkishBrown :: AlphaColour Double

pooBrown :: AlphaColour Double

brightSkyBlue :: AlphaColour Double

celery :: AlphaColour Double

dirtBrown :: AlphaColour Double

strawberry :: AlphaColour Double

darkLime :: AlphaColour Double

copper :: AlphaColour Double

mediumBrown :: AlphaColour Double

mutedGreen :: AlphaColour Double

robin'sEgg :: AlphaColour Double

brightAqua :: AlphaColour Double

brightLavender :: AlphaColour Double

ivory :: AlphaColour Double

veryLightPurple :: AlphaColour Double

lightNavy :: AlphaColour Double

pinkRed :: AlphaColour Double

oliveBrown :: AlphaColour Double

poopBrown :: AlphaColour Double

mustardGreen :: AlphaColour Double

oceanGreen :: AlphaColour Double

veryDarkBlue :: AlphaColour Double

dustyGreen :: AlphaColour Double

lightNavyBlue :: AlphaColour Double

mintyGreen :: AlphaColour Double

adobe :: AlphaColour Double

barney :: AlphaColour Double

jadeGreen :: AlphaColour Double

brightLightBlue :: AlphaColour Double

lightLime :: AlphaColour Double

darkKhaki :: AlphaColour Double

orangeYellow :: AlphaColour Double

ocre :: AlphaColour Double

maize :: AlphaColour Double

fadedPink :: AlphaColour Double

britishRacingGreen :: AlphaColour Double

sandstone :: AlphaColour Double

mudBrown :: AlphaColour Double

lightSeaGreen :: AlphaColour Double

robinEggBlue :: AlphaColour Double

aquaMarine :: AlphaColour Double

darkSeaGreen :: AlphaColour Double

softPink :: AlphaColour Double

orangeyBrown :: AlphaColour Double

cherryRed :: AlphaColour Double

burntYellow :: AlphaColour Double

brownishGrey :: AlphaColour Double

camel :: AlphaColour Double

purplishGrey :: AlphaColour Double

marine :: AlphaColour Double

greyishPink :: AlphaColour Double

paleTurquoise :: AlphaColour Double

pastelYellow :: AlphaColour Double

blueyPurple :: AlphaColour Double

canaryYellow :: AlphaColour Double

fadedRed :: AlphaColour Double

sepia :: AlphaColour Double

coffee :: AlphaColour Double

brightMagenta :: AlphaColour Double

mocha :: AlphaColour Double

ecru :: AlphaColour Double

purpleish :: AlphaColour Double

cranberry :: AlphaColour Double

darkishGreen :: AlphaColour Double

brownOrange :: AlphaColour Double

duskyRose :: AlphaColour Double

melon :: AlphaColour Double

sicklyGreen :: AlphaColour Double

silver :: AlphaColour Double

purplyBlue :: AlphaColour Double

purpleishBlue :: AlphaColour Double

hospitalGreen :: AlphaColour Double

shitBrown :: AlphaColour Double

midBlue :: AlphaColour Double

amber :: AlphaColour Double

easterGreen :: AlphaColour Double

softBlue :: AlphaColour Double

ceruleanBlue :: AlphaColour Double

goldenBrown :: AlphaColour Double

brightTurquoise :: AlphaColour Double

redPink :: AlphaColour Double

redPurple :: AlphaColour Double

greyishBrown :: AlphaColour Double

vermillion :: AlphaColour Double

russet :: AlphaColour Double

steelGrey :: AlphaColour Double

lighterPurple :: AlphaColour Double

brightViolet :: AlphaColour Double

prussianBlue :: AlphaColour Double

slateGreen :: AlphaColour Double

dirtyPink :: AlphaColour Double

darkBlueGreen :: AlphaColour Double

pine :: AlphaColour Double

yellowyGreen :: AlphaColour Double

darkGold :: AlphaColour Double

bluish :: AlphaColour Double

darkishBlue :: AlphaColour Double

dullRed :: AlphaColour Double

pinkyRed :: AlphaColour Double

bronze :: AlphaColour Double

paleTeal :: AlphaColour Double

militaryGreen :: AlphaColour Double

barbiePink :: AlphaColour Double

bubblegumPink :: AlphaColour Double

peaSoupGreen :: AlphaColour Double

darkMustard :: AlphaColour Double

shit :: AlphaColour Double

mediumPurple :: AlphaColour Double

veryDarkGreen :: AlphaColour Double

dirt :: AlphaColour Double

duskyPink :: AlphaColour Double

redViolet :: AlphaColour Double

lemonYellow :: AlphaColour Double

pistachio :: AlphaColour Double

dullYellow :: AlphaColour Double

darkLimeGreen :: AlphaColour Double

denimBlue :: AlphaColour Double

tealBlue :: AlphaColour Double

lightishBlue :: AlphaColour Double

purpleyBlue :: AlphaColour Double

lightIndigo :: AlphaColour Double

swampGreen :: AlphaColour Double

brownGreen :: AlphaColour Double

darkMaroon :: AlphaColour Double

hotPurple :: AlphaColour Double

darkForestGreen :: AlphaColour Double

fadedBlue :: AlphaColour Double

drabGreen :: AlphaColour Double

lightLimeGreen :: AlphaColour Double

snotGreen :: AlphaColour Double

yellowish :: AlphaColour Double

lightBlueGreen :: AlphaColour Double

bordeaux :: AlphaColour Double

lightMauve :: AlphaColour Double

ocean :: AlphaColour Double

marigold :: AlphaColour Double

muddyGreen :: AlphaColour Double

dullOrange :: AlphaColour Double

steel :: AlphaColour Double

electricPurple :: AlphaColour Double

fluorescentGreen :: AlphaColour Double

yellowishBrown :: AlphaColour Double

blush :: AlphaColour Double

softGreen :: AlphaColour Double

brightOrange :: AlphaColour Double

lemon :: AlphaColour Double

purpleGrey :: AlphaColour Double

acidGreen :: AlphaColour Double

paleLavender :: AlphaColour Double

violetBlue :: AlphaColour Double

lightForestGreen :: AlphaColour Double

burntRed :: AlphaColour Double

khakiGreen :: AlphaColour Double

cerise :: AlphaColour Double

fadedPurple :: AlphaColour Double

apricot :: AlphaColour Double

darkOliveGreen :: AlphaColour Double

greyBrown :: AlphaColour Double

greenGrey :: AlphaColour Double

trueBlue :: AlphaColour Double

paleViolet :: AlphaColour Double

periwinkleBlue :: AlphaColour Double

lightSkyBlue :: AlphaColour Double

blurple :: AlphaColour Double

greenBrown :: AlphaColour Double

bluegreen :: AlphaColour Double

brightTeal :: AlphaColour Double

brownishYellow :: AlphaColour Double

peaSoup :: AlphaColour Double

forest :: AlphaColour Double

barneyPurple :: AlphaColour Double

ultramarine :: AlphaColour Double

purplish :: AlphaColour Double

pukeYellow :: AlphaColour Double

bluishGrey :: AlphaColour Double

darkPeriwinkle :: AlphaColour Double

darkLilac :: AlphaColour Double

reddish :: AlphaColour Double

lightMaroon :: AlphaColour Double

dustyPurple :: AlphaColour Double

terraCotta :: AlphaColour Double

avocado :: AlphaColour Double

marineBlue :: AlphaColour Double

tealGreen :: AlphaColour Double

slateGrey :: AlphaColour Double

lighterGreen :: AlphaColour Double

electricGreen :: AlphaColour Double

dustyBlue :: AlphaColour Double

goldenYellow :: AlphaColour Double

brightYellow :: AlphaColour Double

lightLavender :: AlphaColour Double

umber :: AlphaColour Double

poop :: AlphaColour Double

darkPeach :: AlphaColour Double

jungleGreen :: AlphaColour Double

eggshell :: AlphaColour Double

denim :: AlphaColour Double

yellowBrown :: AlphaColour Double

dullPurple :: AlphaColour Double

chocolateBrown :: AlphaColour Double

wineRed :: AlphaColour Double

neonBlue :: AlphaColour Double

dirtyGreen :: AlphaColour Double

lightTan :: AlphaColour Double

iceBlue :: AlphaColour Double

cadetBlue :: AlphaColour Double

darkMauve :: AlphaColour Double

veryLightBlue :: AlphaColour Double

greyPurple :: AlphaColour Double

pastelPink :: AlphaColour Double

veryLightGreen :: AlphaColour Double

darkSkyBlue :: AlphaColour Double

evergreen :: AlphaColour Double

dullPink :: AlphaColour Double

aubergine :: AlphaColour Double

mahogany :: AlphaColour Double

reddishOrange :: AlphaColour Double

deepGreen :: AlphaColour Double

vomitGreen :: AlphaColour Double

purplePink :: AlphaColour Double

dustyPink :: AlphaColour Double

fadedGreen :: AlphaColour Double

camoGreen :: AlphaColour Double

pinkyPurple :: AlphaColour Double

pinkPurple :: AlphaColour Double

brownishRed :: AlphaColour Double

darkRose :: AlphaColour Double

mud :: AlphaColour Double

brownish :: AlphaColour Double

emeraldGreen :: AlphaColour Double

paleBrown :: AlphaColour Double

dullBlue :: AlphaColour Double

burntUmber :: AlphaColour Double

mediumGreen :: AlphaColour Double

clay :: AlphaColour Double

lightAqua :: AlphaColour Double

lightOliveGreen :: AlphaColour Double

brownishOrange :: AlphaColour Double

darkAqua :: AlphaColour Double

purplishPink :: AlphaColour Double

darkSalmon :: AlphaColour Double

greenishGrey :: AlphaColour Double

jade :: AlphaColour Double

uglyGreen :: AlphaColour Double

darkBeige :: AlphaColour Double

emerald :: AlphaColour Double

paleRed :: AlphaColour Double

lightMagenta :: AlphaColour Double

sky :: AlphaColour Double

lightCyan :: AlphaColour Double

yellowOrange :: AlphaColour Double

reddishPurple :: AlphaColour Double

reddishPink :: AlphaColour Double

orchid :: AlphaColour Double

dirtyYellow :: AlphaColour Double

orangeRed :: AlphaColour Double

deepRed :: AlphaColour Double

orangeBrown :: AlphaColour Double

cobaltBlue :: AlphaColour Double

neonPink :: AlphaColour Double

rosePink :: AlphaColour Double

greyishPurple :: AlphaColour Double

raspberry :: AlphaColour Double

aquaGreen :: AlphaColour Double

salmonPink :: AlphaColour Double

tangerine :: AlphaColour Double

brownishGreen :: AlphaColour Double

redBrown :: AlphaColour Double

greenishBrown :: AlphaColour Double

pumpkin :: AlphaColour Double

pineGreen :: AlphaColour Double

charcoal :: AlphaColour Double

babyPink :: AlphaColour Double

cornflower :: AlphaColour Double

blueViolet :: AlphaColour Double

chocolate :: AlphaColour Double

greyishGreen :: AlphaColour Double

scarlet :: AlphaColour Double

greenYellow :: AlphaColour Double

darkOlive :: AlphaColour Double

sienna :: AlphaColour Double

pastelPurple :: AlphaColour Double

terracotta :: AlphaColour Double

aquaBlue :: AlphaColour Double

sageGreen :: AlphaColour Double

bloodRed :: AlphaColour Double

deepPink :: AlphaColour Double

grass :: AlphaColour Double

moss :: AlphaColour Double

pastelBlue :: AlphaColour Double

bluishGreen :: AlphaColour Double

greenBlue :: AlphaColour Double

darkTan :: AlphaColour Double

greenishBlue :: AlphaColour Double

paleOrange :: AlphaColour Double

vomit :: AlphaColour Double

forrestGreen :: AlphaColour Double

darkLavender :: AlphaColour Double

darkViolet :: AlphaColour Double

purpleBlue :: AlphaColour Double

darkCyan :: AlphaColour Double

oliveDrab :: AlphaColour Double

pinkish :: AlphaColour Double

cobalt :: AlphaColour Double

neonPurple :: AlphaColour Double

lightTurquoise :: AlphaColour Double

appleGreen :: AlphaColour Double

dullGreen :: AlphaColour Double

wine :: AlphaColour Double

powderBlue :: AlphaColour Double

offWhite :: AlphaColour Double

electricBlue :: AlphaColour Double

darkTurquoise :: AlphaColour Double

bluePurple :: AlphaColour Double

azure :: AlphaColour Double

brightRed :: AlphaColour Double

pinkishRed :: AlphaColour Double

cornflowerBlue :: AlphaColour Double

lightOlive :: AlphaColour Double

grape :: AlphaColour Double

greyishBlue :: AlphaColour Double

purplishBlue :: AlphaColour Double

yellowishGreen :: AlphaColour Double

greenishYellow :: AlphaColour Double

mediumBlue :: AlphaColour Double

dustyRose :: AlphaColour Double

lightViolet :: AlphaColour Double

midnightBlue :: AlphaColour Double

bluishPurple :: AlphaColour Double

redOrange :: AlphaColour Double

darkMagenta :: AlphaColour Double

greenish :: AlphaColour Double

oceanBlue :: AlphaColour Double

coral :: AlphaColour Double

cream :: AlphaColour Double

reddishBrown :: AlphaColour Double

burntSienna :: AlphaColour Double

brick :: AlphaColour Double

sage :: AlphaColour Double

greyGreen :: AlphaColour Double

white :: AlphaColour Double

robin'sEggBlue :: AlphaColour Double

mossGreen :: AlphaColour Double

steelBlue :: AlphaColour Double

eggplant :: AlphaColour Double

lightYellow :: AlphaColour Double

leafGreen :: AlphaColour Double

lightGrey :: AlphaColour Double

puke :: AlphaColour Double

pinkishPurple :: AlphaColour Double

seaBlue :: AlphaColour Double

palePurple :: AlphaColour Double

slateBlue :: AlphaColour Double

blueGrey :: AlphaColour Double

hunterGreen :: AlphaColour Double

fuchsia :: AlphaColour Double

crimson :: AlphaColour Double

paleYellow :: AlphaColour Double

ochre :: AlphaColour Double

mustardYellow :: AlphaColour Double

lightRed :: AlphaColour Double

cerulean :: AlphaColour Double

palePink :: AlphaColour Double

deepBlue :: AlphaColour Double

rust :: AlphaColour Double

lightTeal :: AlphaColour Double

slate :: AlphaColour Double

goldenrod :: AlphaColour Double

darkYellow :: AlphaColour Double

darkGrey :: AlphaColour Double

armyGreen :: AlphaColour Double

greyBlue :: AlphaColour Double

seafoam :: AlphaColour Double

puce :: AlphaColour Double

springGreen :: AlphaColour Double

darkOrange :: AlphaColour Double

sand :: AlphaColour Double

pastelGreen :: AlphaColour Double

mint :: AlphaColour Double

lightOrange :: AlphaColour Double

brightPink :: AlphaColour Double

chartreuse :: AlphaColour Double

deepPurple :: AlphaColour Double

darkBrown :: AlphaColour Double

taupe :: AlphaColour Double

peaGreen :: AlphaColour Double

pukeGreen :: AlphaColour Double

kellyGreen :: AlphaColour Double

seafoamGreen :: AlphaColour Double

blueGreen :: AlphaColour Double

khaki :: AlphaColour Double

burgundy :: AlphaColour Double

darkTeal :: AlphaColour Double

brickRed :: AlphaColour Double

royalPurple :: AlphaColour Double

plum :: AlphaColour Double

mintGreen :: AlphaColour Double

gold :: AlphaColour Double

babyBlue :: AlphaColour Double

yellowGreen :: AlphaColour Double

brightPurple :: AlphaColour Double

darkRed :: AlphaColour Double

paleBlue :: AlphaColour Double

grassGreen :: AlphaColour Double

navy :: AlphaColour Double

aquamarine :: AlphaColour Double

burntOrange :: AlphaColour Double

neonGreen :: AlphaColour Double

brightBlue :: AlphaColour Double

rose :: AlphaColour Double

lightPink :: AlphaColour Double

mustard :: AlphaColour Double

indigo :: AlphaColour Double

lime :: AlphaColour Double

seaGreen :: AlphaColour Double

periwinkle :: AlphaColour Double

darkPink :: AlphaColour Double

oliveGreen :: AlphaColour Double

peach :: AlphaColour Double

paleGreen :: AlphaColour Double

lightBrown :: AlphaColour Double

hotPink :: AlphaColour Double

black :: AlphaColour Double

lilac :: AlphaColour Double

navyBlue :: AlphaColour Double

royalBlue :: AlphaColour Double

beige :: AlphaColour Double

salmon :: AlphaColour Double

olive :: AlphaColour Double

maroon :: AlphaColour Double

brightGreen :: AlphaColour Double

darkPurple :: AlphaColour Double

mauve :: AlphaColour Double

forestGreen :: AlphaColour Double

aqua :: AlphaColour Double

cyan :: AlphaColour Double

tan :: AlphaColour Double

darkBlue :: AlphaColour Double

lavender :: AlphaColour Double

turquoise :: AlphaColour Double

darkGreen :: AlphaColour Double

violet :: AlphaColour Double

lightPurple :: AlphaColour Double

limeGreen :: AlphaColour Double

grey :: AlphaColour Double

skyBlue :: AlphaColour Double

yellow :: AlphaColour Double

magenta :: AlphaColour Double

lightGreen :: AlphaColour Double

orange :: AlphaColour Double

teal :: AlphaColour Double

lightBlue :: AlphaColour Double

red :: AlphaColour Double

brown :: AlphaColour Double

pink :: AlphaColour Double

blue :: AlphaColour Double

green :: AlphaColour Double

purple :: AlphaColour Double

xkcdColorMap :: Map String (AlphaColour Double)