Diagrams.Lens

Description

This module provides utilities for using Control.Lens with diagrams, including orphan instances for the Wrapped class.

Documentation

type Wrapped' s a = Wrapped s s a a

_P :: Iso s t (Point s) (Point t)

Diagrams.Align

_envelopeVMove :: (Monoid a, HasOrigin a, Enveloped a, Num (Scalar (V a))) => V a -> Traversal' a (V a)

A singleton Traversal for an envelope vector, where modification of the vector moves the origin appropriately. No vector is traversed when the envelope is empty.

This is the same as _alignedVMove with 1 as the interpolation parameter.

_alignedVMove :: (Monoid a, HasOrigin a, Enveloped a, Num (Scalar (V a))) => V a -> Scalar (V a) -> Traversal' a (V a)

A singleton Traversal for an alignment vector, where modification of the vector moves the origin appropriately. No vector is traversed when the envelope is empty.

The interface mimics Diagrams.Align.alignBy in that the d parameter specifies an interpolation between two extremes of the envelope. d = 1 is on the envelope along the vector, whereas d = -1 is on the envelope, away from the vector.

If you need a Point instead of a vector, then compose with _P.

Diagrams.BoundingBox

_corners :: (HasBasis v, Ord (Basis v), AdditiveGroup (Scalar v), Ord (Scalar v), HasBasis v', Ord (Basis v'), AdditiveGroup (Scalar v'), Ord (Scalar v')) => Traversal (BoundingBox v) (BoundingBox v') (Point v) (Point v')

A traversal that either has 0 (empty box) or 2 points. These points are the lower and upper corners, respectively.

_boxExtents :: (HasBasis v, Ord (Basis v), AdditiveGroup (Scalar v), Ord (Scalar v)) => Lens' (BoundingBox v) v

A lens that gets the extents of the box. In order to change the extents, this modifies the upper corner.

Diagrams.Core.Style

_attr :: AttributeClass a => Lens' (Style v) (Maybe a)

_mkAttr :: AttributeClass a => Prism' (Attribute v) a

_mkTAttr :: (AttributeClass a, Transformable a, V a ~ v) => Prism' (Attribute v) a

Diagrams.Core.Types

_location :: (HasLinearMap v, InnerSpace v, OrderedField (Scalar v)) => Lens' (Subdiagram b v m) (Point v)

Gets or set the location of a Subdiagram.

_mkSubdiagram :: (HasLinearMap v, InnerSpace v, OrderedField (Scalar v), Semigroup m) => Iso' (QDiagram b v m) (Subdiagram b v m)

Diagrams.Located

_Loc :: Iso (Located a) (Located a') (Point (V a), a) (Point (V a'), a')

Diagrams.Parametric

Diagrams.Segment

_mkFixedSeg :: (AdditiveGroup v, AdditiveGroup v') => Iso (Located (Segment Closed v)) (Located (Segment Closed v')) (FixedSegment v) (FixedSegment v')

_straight :: Prism' (Segment Closed v) v

Prism that constructs linear segments. Can also destruct them, if the segment is Linear.

_bezier3 :: Prism' (Segment Closed v) (v, v, v)

Prism that constructs cubic bezier segments. Can also destruct them, if segment is a Cubic.

Diagrams.Trail

_lineVertices :: (InnerSpace v, OrderedField (Scalar v), InnerSpace v', OrderedField (Scalar v')) => Iso (Located (Trail' Line v)) (Located (Trail' Line v')) [Point v] [Point v']

_lineOffsets :: (InnerSpace v, OrderedField (Scalar v), InnerSpace v', OrderedField (Scalar v')) => Iso (Trail' Line v) (Trail' Line v') [v] [v']

_lineSegments :: (InnerSpace v, OrderedField (Scalar v), InnerSpace v', OrderedField (Scalar v')) => Iso (Trail' Line v) (Trail' Line v') [Segment Closed v] [Segment Closed v']

Diagrams.TwoD.Types

_toTurn :: (Angle a, Angle a') => Iso a a' Turn Turn

toTurn is an isomorphism from angles to Turns.