Diagrams.TwoD.Path.Turtle

Description

Stateful domain specific language for diagram paths, modelled after the classic "turtle" graphics language.

Documentation

type Turtle = TurtleT Identity

type TurtleT = StateT TurtleState

Turtle control commands

runTurtle :: Turtle a -> TurtleState

Run the turtle, yielding the final turtle state.

runTurtleT :: Monad m => TurtleT m a -> m TurtleState

A more general way to run the turtle. Returns a computation in the underlying monad m yielding the final turtle state.

drawTurtle :: Renderable (Path R2) b => Turtle a -> Diagram b R2

Run the turtle, yielding a diagram.

drawTurtleT :: (Monad m, Functor m, Renderable (Path R2) b) => TurtleT m a -> m (Diagram b R2)

A more general way to run the turtle. Returns a computation in the underlying monad m yielding the final diagram.

sketchTurtle :: Turtle a -> Path R2

Run the turtle, ignoring any pen style commands and yielding a 2D path.

sketchTurtleT :: (Functor m, Monad m) => TurtleT m a -> m (Path R2)

A more general way to run the turtle. Returns a computation in the underlying monad m, ignoring any pen style commands and yielding a 2D path.

Motion commands

forward :: Monad m => Double -> TurtleT m ()

Move the turtle forward, along the current heading.

backward :: Monad m => Double -> TurtleT m ()

Move the turtle backward, directly away from the current heading.

left :: Monad m => Double -> TurtleT m ()

Modify the current heading to the left by the specified angle in degrees.

right :: Monad m => Double -> TurtleT m ()

Modify the current heading to the right by the specified angle in degrees.

State accessors / setters

heading :: Monad m => TurtleT m Double

Get the current turtle angle, in degrees.

setHeading :: Monad m => Double -> TurtleT m ()

Set the current turtle angle, in degrees.

towards :: Monad m => P2 -> TurtleT m ()

Sets the heading towards a given location.

isDown :: Monad m => TurtleT m Bool

Queries whether the pen is currently drawing a path or not.

pos :: Monad m => TurtleT m P2

Get the current turtle X/Y position.

setPos :: Monad m => P2 -> TurtleT m ()

Set the current turtle X/Y position.

setPenWidth :: Monad m => Double -> TurtleT m ()

Sets the pen size

setPenColor :: Monad m => Colour Double -> TurtleT m ()

Sets the pen color

Drawing control

penUp :: Monad m => TurtleT m ()

Ends the current path, and enters into penUp mode

penDown :: Monad m => TurtleT m ()

Ends the current path, and enters into penDown mode

penHop :: Monad m => TurtleT m ()

Start a new trail at current position

closeCurrent :: Monad m => TurtleT m ()

Closes the current path , to the starting position of the current trail. Has no effect when the pen position is up.